AI_Mouse_Controller/learning_mode.py
import pyautogui
import time
import logging

class LearningMode:
    def __init__(self):
        self.model = None

    def load_model(self):
        # Load the AI model
        logging.info('Loading AI model...')
        time.sleep(1)
        logging.info('AI model loaded.')

    def record_mouse_movement(self):
        # Record mouse positions
        logging.info('Recording mouse movement...')
        time.sleep(1)
        logging.info('Mouse movement recorded.')

    def train_model(self):
        # Train the model
        logging.info('Training model...')
        time.sleep(1)
        logging.info('Model trained.')

    def move_mouse(self):
        # Move the mouse to the predicted position
        logging.info('Moving mouse...')
        time.sleep(1)
        self.predict_next_position()
        pyautogui.moveTo(500, 500)  # Replace with the predicted position
        logging.info('Mouse moved.')

    def predict_next_position(self):
        # Predict the next mouse position
        logging.info('Predicting next mouse position...')
        time.sleep(1)
        logging.info('Next mouse position predicted.')
