AI_Mouse_Controller/tests/test_learning_mode.py
import unittest
from learning_mode import LearningMode

class TestLearningMode(unittest.TestCase):
    def test_load_model(self):
        learning_mode = LearningMode()
        learning_mode.load_model()
        # Add assertions here

    def test_record_mouse_movement(self):
        learning_mode = LearningMode()
        learning_mode.record_mouse_movement()
        # Add assertions here

    def test_train_model(self):
        learning_mode = LearningMode()
        learning_mode.train_model()
        # Add assertions here

    def test_move_mouse(self):
        learning_mode = LearningMode()
        learning_mode.move_mouse()
        # Add assertions here

if __name__ == '__main__':
    unittest.main()
